requirements.txt
# Usage: pip install -r requirements.txt

# For YOLOv7

# Base ----------------------------------------
matplotlib>=3.2.2
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0

# Logging -------------------------------------
tensorboard>=2.4.1

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation

# For Tennis-Technique
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
Django==5.0.4
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.0
flatbuffers==24.3.25
fsspec==2024.5.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
keras==3.4.1
keras-tuner==1.4.7
kt-legacy==1.0.5
libclang==18.1.1
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mosspy==1.0.9
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==4.25.3
psycopg2==2.9.9
Pygments==2.18.0
requests==2.32.3
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==70.3.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.4.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0